kiosk assignments, and correct location for checkin/out
