+++ src/ChurchCRM/model/ChurchCRM/Event.php
@@ -3,7 +3,7 @@
 namespace ChurchCRM;
 
 use ChurchCRM\Base\Event as BaseEvent;
-
+use Propel\Runtime\ActiveQuery\Criteria;
 /**
  * Skeleton subclass for representing a row from the 'events_event' table.
  *
@@ -15,4 +15,38 @@ use ChurchCRM\Base\Event as BaseEvent;
  */
 class Event extends BaseEvent
 {
+  
+  public function checkInPerson($PersonId)
+  {    
+    $AttendanceRecord = EventAttendQuery::create()
+            ->filterByEvent($this)
+            ->filterByPersonId($PersonId)
+            ->findOneOrCreate();
+    
+    $AttendanceRecord->setEvent($this)
+      ->setPersonId($PersonId)
+      ->setCheckinDate(date('Y-m-d H:i:s'))
+      ->setCheckoutDate(null)
+      ->save();
+    
+    return array("status"=>"success");
+    
+  }
+  
+  public function checkOutPerson($PersonId)
+  {    
+    $AttendanceRecord = EventAttendQuery::create()
+            ->filterByEvent($this)
+            ->filterByPersonId($PersonId)
+            ->filterByCheckinDate(NULL,  Criteria::NOT_EQUAL)
+            ->findOne();
+    
+    $AttendanceRecord->setEvent($this)
+      ->setPersonId($PersonId)
+      ->setCheckoutDate(date('Y-m-d H:i:s'))
+      ->save();
+    
+    return array("status"=>"success");
+    
+  }
 }
