+++ src/ChurchCRM/model/ChurchCRM/KioskAssignmentQuery.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace ChurchCRM;
+
+use ChurchCRM\Base\KioskAssignmentQuery as BaseKioskAssignmentQuery;
+
+/**
+ * Skeleton subclass for performing query and update operations on the 'kioskassginment_kasm' table.
+ *
+ *
+ *
+ * You should add additional methods to this class to meet the
+ * application requirements.  This class will only be generated as
+ * long as it does not already exist in the output directory.
+ *
+ */
+class KioskAssignmentQuery extends BaseKioskAssignmentQuery
+{
+
+}
